Disable input fields for Modals in AddPatient
When in the Add Patient Page, users are still able to interact with the
inputs for the other categories aside from Admin, as well as press the
save button. This should not be possible since according to the desired
workflow, users will only edit the fields after the Admin category for a
patient has been created.

Disabled all the inputs for the other modals, as well as hid /
removed the edit and save buttons when in Add Patient mode.