Add support for adding visits
AddNewVisitModal is similar to Admin modal in terms of fields, but does
not have isAdd or isEditing props to decide whether its fields are
disabled.

It can only take in admin fields of a new visit. After successful
creation of a new visit, the page will be redirected to the new visit of
the patient.

Sidebar now takes an isAdd prop that decides whether to display the Add
New Visit button.