Add Fallrisk Category
Fallrisk category is for patients above 60 years of age, and comprises a few multiple option questions, with each
option associated with a score.
A score of >=8 is high risk.
The backend does not store the mappings of scores and question
information, only the selected alphabets for each question
(a,b,c,d,e...). The client does the calculations.